Fix nested lists in getParent (renamed to getParentNode)
Explicitly support the concept of having multiple indexes in a row.

Note that the old code's use of getLevel() was not correct: that undocumented
method means "how many named fields are in the entire path", which isn't helpful
here.

The test change is because the case in question was actually being tested, but
the test was wrong.