add support for graphql-java context map
`graphql-java` v17 introduced new `GraphQLContext` mechanism which standardized how context should be utilized. Prior to v17, context could be of any object and it was up to the developers to decide how it should look like. New `GraphQLContext` standardizes the approach and is a simple wrapper around a map.

Updated default `FederatedTracingInstrumentation.Options#shouldTrace` method to also check the `GraphQLContext` map contents if it contains an entry for `apollo-federation-include-trace`.