update SDL generation logic to work with applied directives (#231)
`FederationSdlPrinter` is a copy of a `graphql.schema.idl.SchemaPrinter` from `graphql-java` v17 with some custom filtering logic that was required for Federation v1.

`graphql-java` v18 introduced concept of applied directive to make a distinction from the directive definition (previously same `GraphQLDirective` type was used for both). This PR drops our custom copy of the schema printer and instead updates our logic to rely on the built-in schema printer provided by the `graphql-java`. Using custom predicates for filtering directives and schema elements we can replicate our custom functionality without the need of a custom printer.

Related:

* resolves: https://github.com/apollographql/federation-jvm/issues/227
* resolves: https://github.com/apollographql/federation-jvm/issues/216
* supersedes: https://github.com/apollographql/federation-jvm/pull/223