fix Import scalar namespace
`@link` directive can be used to import various definitions by specifying `import: [Import]` parameter. `Import` is custom scalar that can either be a String or an Object (that allows to rename the imported type).

`Import` scalar was imported with incorrect namespace as `federation__Import` vs the expected `link__Import`. Updating transformation logic to correctly namespace the scalar. Namespace can be skipped if `Import` scalar is imported from the `@link` spec.