ensure scalars are wired up if no existing wiring (#382)
Fixes #380

This restructures the scalar transformation logic to ensure that both
the type definition and wiring are added correctly if necessary. This
allows consumers to define the scalars in their schema, but not assign
runtime wiring. This can be nice to support other tools like linters or
graphql-faker that require the schema to be a valid schema file.