downgrade protobuf dependency to v3 (#428)
We are using protobuf to generate `ftv1` tracing information. Protobuf
v4 made a backwards incompatible change that prevents using those
messages with v3 services. Since there are a lot of existing libraries
(including `grpc-java`) that still depend on the v3 format and generated
protobuf v3 message can be used for both v3 and v4 services we'll be
reverting to v3 as well.

See:
https://github.com/grpc/grpc-java/issues/11015#issuecomment-2560196695

Resolves: https://github.com/apollographql/federation-jvm/issues/421